Fix build on Ubuntu 23.04 (kernel 6.2).
This is done by specifying the ranges for 6.3 and 6.4 versions which need the fix.
Kernel 6.2, to my knowledge, is not used in other popular distros than Ubuntu.

Ubuntu had backported the change from kernel 6.5 to 6.2, which is why this is needed.

Added a comment instructing what to do if one wants to build the module
running kernel 6.2 on something else than Ubuntu.