allow up to 4052 byte packets (#31)
Tested on a pair of RTL8812AUs over USB2 on two different systems
sending max size packets and receiving and verifying each packet. No
kernel panics even on a heavily loaded system, whereas previously it
was easy to trigger a panic or a rejection of the packet.

Seems to work fine even without the receive side change, but better
safe than sorry.

Designed to work with max 4000 bytes of user data with wfb-ng.